Randomness

+++ event_simulation/src/main/java/FileOperations/InputProcessor.java
@@ -8,13 +8,13 @@ import java.util.*;
 
 public class InputProcessor {
 
-    public static  HashMap<Integer, float[][]> readSignalData(String filePath) throws IOException{
+    public static  HashMap<Integer, double[][]> readSignalData(String filePath) throws IOException{
         File file = new File(filePath);
         BufferedReader br = new BufferedReader(new FileReader(file));
         String line_str;
         String line[] = new String[20];
         int intersection_num =0;
-        HashMap<Integer, float[][]> signalMap = new HashMap<Integer, float[][]>();
+        HashMap<Integer, double[][]> signalMap = new HashMap<Integer, double[][]>();
         /*
         * 1,[[8	1.8	1.8	30	3.8	55], [5	3.6	4.2	28	3.8	55], []]
         * 2, [[][][]]
@@ -28,28 +28,28 @@ public class InputProcessor {
             if(intervals[0].startsWith("Peachtree")){
                 intersection_num++;
                 if(intersection_num==3){ // 13th st. data not given
-                    signalMap.put(intersection_num, new float[3][8]);
+                    signalMap.put(intersection_num, new double[3][8]);
                     intersection_num++;
                 }
-                signalMap.put(intersection_num, new float[3][8]);
+                signalMap.put(intersection_num, new double[3][8]);
             }else if(intervals[0].equals("Eastbound") || intervals[0].equals("Westbound") || intervals[0].equals("Northbound")){
-                float[][] lights = signalMap.get(intersection_num);
+                double[][] lights = signalMap.get(intersection_num);
                 int direction;
-                int totalTurnTime=0;
-                int totalTRTime =0;
+                double totalTurnTime=0.00;
+                double totalTRTime =0.00;
                 if(intervals[0].equals("Eastbound"))direction=0;
                 else if(intervals[0].equals("Northbound")) direction=1;
                 else direction=2;
                 for(int i=0; i<3;i++){
                     if(intervals[i+1].isEmpty())
                         intervals[i+1]="0";
-                    lights[direction][i] = Float.parseFloat(intervals[i+1]);
+                    lights[direction][i] = Double.parseDouble(intervals[i+1]);
                     totalTurnTime+=lights[direction][i];
                 }
                 for(int i=3; i<6;i++){
                     if(intervals[i+1].isEmpty())
                         intervals[i+1]="0";
-                    lights[direction][i] = Float.parseFloat(intervals[i+1]);
+                    lights[direction][i] = Double.parseDouble(intervals[i+1]);
                     totalTRTime+=lights[direction][i];
                 }
                 lights[direction][6]=totalTurnTime;
